Day 2
It was easy. With small steps, I am going to become a Data Scientist. 
I believe it will be an amazing trip. I really want to achieve an expertise degree in this area.
Maybe it is a crucial move in my life. I hope so!
So, Don't waste time! Let's jump in for coding!!!